pyproject.toml
[project]
name = "bbgregressions"
version = "0.1.0"
description = "Customized statistical regressions analysis"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Barcelona Biomedical Genomics Lab", email = "bbglab@irbbarcelona.org"}]
maintainers = [
    {name = "Raquel Blanco", email = "raquel.blanco@irbbarcelona.org"}
]
keywords = ["bbgregressions", "deepcsa"]


requires-python = ">=3.13"
dependencies = [
    "click>=8.3.1",
    "daiquiri>=3.4.0",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.5",
]


[project.scripts]
bbgregressions = "src.main:bbgregressions" 

[tool.uv]
package = true

[build-system]
requires = ["hatchling", "cython==0.29.36"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["src"]